add strategist vault trait impl; rename bridge trait to routing
