+++ e2e/examples/eth_vault/strategist/mod.rs
@@ -1,3 +1,4 @@
 mod astroport;
-mod bridge;
 pub mod client;
+mod routing;
+mod vault;
